Use $USER as bootstrap users
This reverts to the default value of 'controller_bootstrap_user',
'seed_bootstrap_user' and 'seed_hypervisor_bootstrap_user' prior to the
addition of the os_distribution variable in Kayobe
(https://review.opendev.org/c/openstack/kayobe/+/787352). In an
all-in-one environment, using the current user as the bootstrap user
makes sense.

Story: 2004960
Task: 42386

Change-Id: I814cd3d97e850a9b0062f92ec5622a756bed0347