set cpu_model to Nehalem for kvm jobs
RHEL 9 and Ubuntu Jammy are being compiled for the x86_64-v2
architecture which is newer than the qemu default of qemu64.
Nehalem is apparently the oldest model that works for x86_64-v2
and is expected to work on Intel and AMD cpus with kvm or qemu.

See devstack change [0].

[0]: Ibd6e11b59f3c8655bc60ace7383a08458b2177f2

Depends-on: https://review.opendev.org/c/openstack/kayobe/+/872981

Change-Id: I09567b4cce95ff1fe2fe4e9e91e49d5f5082f7d0
(cherry picked from commit 416da1f522753fefe2559e6224cb0c50cb177fe5)