Increase resource capacity for lab controller
Our standard lab resources are bigger than they once were.  When
running a lab spread over several weeks, it is a frequent occurrence
for the controller node to run out of disk or RAM.  This leads to
oom-killer, filesystem overflow and other instability we can do
without.  Increase both here.