Set explicit MTU on dummy/bridge lab network connections
This resolves the issue where on NetworkManager systems (RL9) the
seed hypervisor host configure fails on a check on braio MTU.  The
issue appears to be that a pre-existing braio is created by
configure-local-networking.sh that has incorrect MTU assignment.
The configuration applied by MichaelRigart.interfaces creates a
second configuration set for braio, but the existing configuration
takes precedence.

It may be better to use nmcli to create the braio and dummy1
interfaces in the first place.