Use dummy1 as bridge port instead of eth1
With NetworkManager, the MichaelRigart.interfaces role only creates an
interface with the dummy type if it is called dummy*, otherwise it is
created as ethernet.

The default dev configuration was using eth1 for the bridge port, which
was causing issues on Rocky Linux 9, for example when following the
automated setup [1]. Switch to dummy1 to ensure a dummy port is created.

[1] https://docs.openstack.org/kayobe/latest/contributor/automated.html

Change-Id: Ib61a62e126c66d4133d8e80dff5643ebc6966bd1