Update to latest cloud images
We were pinning to old cloud images, but we should now be able to switch
to the latest ones since we use UEFI by default [1].

A recent Bifrost change [2] introduced various improvements surrounding
deployment images that are downloaded rather than built locally. Bifrost
is now aware of how to download official images, so we only need to
specify the Linux distribution and release.

However, there is a bug affecting centos/rocky which are using a
different file format for checksums [3]. Disable checksum verification
for these distributions until it is fixed.

[1] https://review.opendev.org/c/openstack/kayobe/+/927015
[2] https://review.opendev.org/c/openstack/bifrost/+/884888
[3] https://bugs.launchpad.net/bifrost/+bug/2081031